requirements.txt

Django==5.0.6
django-ckeditor==6.7.1
django-grappelli==4.0.1
djangorestframework==3.15.1
psycopg2-binary==2.9.9
requests==2.32.3
tzdata==2024.1
